implementation made

+++ src/Service/MyTraclist/MyTraclistService.php
@@ -5,6 +5,8 @@ use App\Interfaces\MyTracklist\MyTracklistInterface;
 use Doctrine\ORM\EntityManagerInterface;
 use App\Entity\Track;
 use App\Service\FileUploader;
+use phpDocumentor\Reflection\Types\Null_;
+use Symfony\Component\Filesystem\Filesystem;
 
 class MyTraclistService implements MyTracklistInterface
 {
@@ -62,19 +64,15 @@ class MyTraclistService implements MyTracklistInterface
         }
         if($tracklistDTO->cover !== NULL)
         {
-            $track->setCover($this->fileUploader->getTargetDirectory()."/". $this->fileUploader->upload($tracklistDTO->cover));
+            $track->setCover($this->fileUploader->upload($tracklistDTO->cover)->getUrl());
         }
 
-        //here we take target directory and make upload file which return string of file URL 
-        $track->setTrackPath($this->fileUploader->getTargetDirectory()."/". $this->fileUploader->upload($tracklistDTO->track_path));
+        $track->setTrackPath($this->fileUploader->upload($tracklistDTO->track_path)->getUrl());
  
         $track->setTitle($tracklistDTO->title);
         $track->setAuthor($tracklistDTO->author);
         $track->setType($tracklistDTO->type);
         $track->setGenre($tracklistDTO->genre);
-
-        $track->setCreatedAt(new \DateTimeImmutable());
-        $track->setUpdatedAt(new \DateTimeImmutable());
 
         $this->entityManager->persist($track);
         $this->entityManager->flush();
@@ -82,23 +80,84 @@ class MyTraclistService implements MyTracklistInterface
         return $track;
     }
 
-    public function showService()
-    {
+    public function showService($id)
+    {   
+
+        if ($this->trackRepository->find($id) == NULL ) 
+        {
+                        return array([
+                'success' => false,
+                'body'    => 'Can not find track'
+            ]); 
+        }
+        return $this->trackRepository->find($id);
 
     }
 
-    public function editService()
+    public function editService($id)
     {
+
+        if ($this->trackRepository->find($id) == NULL ) 
+        {
+                        return array([
+                'success' => false,
+                'body'    => 'Can not find track'
+            ]); 
+        }
+        return $this->trackRepository->find($id);
 
     }
 
-    public function updateService()
+    public function updateService($tracklistDTO, $track)
     {
+        $fileSystem = new Filesystem;
 
+        if($tracklistDTO->cover !== NULL)
+        {
+            $fileSystem->remove('../public/uploads/'.$track->getCover());
+            $track->setCover($this->fileUploader->upload($tracklistDTO->cover)->getUrl());
+        }
+
+        if($tracklistDTO->album !== NULL)
+        {
+            $track->setAlbum($tracklistDTO->album);
+        }
+
+        if($tracklistDTO->title !== NULL)
+        {
+            $track->setTitle($tracklistDTO->title);
+        }
+
+        if ($tracklistDTO->author !== NULL) 
+        {
+            $track->setAuthor($tracklistDTO->author);
+        }
+
+        if ($tracklistDTO->type !== NULL) 
+        {
+            $track->setType($tracklistDTO->type);
+        }
+
+        if ($tracklistDTO->genre !== NULL) 
+        {
+            $track->setGenre($tracklistDTO->genre);
+        }
+        
+        $this->entityManager->flush();
+      
+        return $track;
     }
 
     public function deleteService($id)
     {
+        if ($this->trackRepository->find($id) == NULL ) 
+        {
+            return array([
+                'success' => false,
+                'body'    => 'Can not find track'
+            ]);
+        }
+        
         $this->entityManager->remove($this->trackRepository->find($id));
         $this->entityManager->flush();
